fixed error with the noise pointer that wasn't correctly copied back to global memory

+++ mppi_gpu/include/cost.hpp
@@ -4,22 +4,22 @@
 class Cost{
 public:
     __host__ __device__ Cost();
-    
+
     __host__ __device__ Cost(float* w,
-                             size_t w_size,
+                             int w_size,
                              float* goal,
-                             size_t goal_size,
+                             int goal_size,
                              float lambda,
                              float* inv_s,
-                             size_t u_size);
+                             int u_size);
 
     __host__ __device__ void init(float* w,
-                                  size_t w_size,
+                                  int w_size,
                                   float* goal,
-                                  size_t goal_size,
+                                  int goal_size,
                                   float lambda,
                                   float* inv_s,
-                                  size_t u_size);
+                                  int u_size);
 
     __host__ __device__ float step_cost(float* x, float* u, float* e);
     __host__ __device__ float final_cost(float* x);
